move formatAnsibleMetaData.ts inside utils directory
